Переносит загрузку фильмов в компонент App
