+++ src/components/loading-screen/loading-screen.jsx
@@ -2,7 +2,16 @@ import React from 'react';
 
 const LoadingScreen = () => {
   return (
-    <p>Loading ...</p>
+    <div className="user-page">
+      <div className="user-page__content page-not-found" style={
+        {
+          paddingTop: `100px`,
+          textAlign: `center`,
+        }
+      }>
+        <h1>Loading...</h1>
+      </div>
+    </div>
   );
 };
